# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AfyaMind telemedicine platform REST API.\n**Authentication:** Most protected endpoints require a JWT Bearer token.\nLogin via POST /auth/login, then click the **Authorize** button above and enter: `Bearer \u003cyour_token\u003e`",
        "title": "AfyaMind API",
        "contact": {
            "name": "AfyaMind Support",
            "email": "support@afyamind.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/clinics/{clinicId}/invitations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists all sent invitations for a specific clinic. Super Admin or Clinic Admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clinics"
                ],
                "summary": "List invitations in a clinic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic ID",
                        "name": "clinicId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of invitations",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.DataEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/clinics.InvitationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends an invitation email to a doctor. Accessible by clinic admins.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitations"
                ],
                "summary": "Invite a doctor to the clinic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clinic UUID",
                        "name": "clinicId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Email payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invitations.CreateInvitationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Invitation created successfully",
                        "schema": {
                            "$ref": "#/definitions/response.MessageEnvelope"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Unauthorized to invite",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/encounters/{id}/diagnoses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all diagnoses for a specific encounter.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Get diagnoses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of diagnoses",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/diagnoses.Diagnosis"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Encounter not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a new diagnosis to a specific encounter.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Add a diagnosis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Encounter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Diagnosis details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/diagnoses.CreateDiagnosisRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Diagnosis created successfully",
                        "schema": {
                            "$ref": "#/definitions/diagnoses.Diagnosis"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Encounter not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Encounter is closed",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/invitations/{token}/accept": {
            "post": {
                "description": "Doctor accepts invitation and fills in profile to complete registration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitations"
                ],
                "summary": "Accept invitation \u0026 complete signup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation Token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sign up details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invitations.AcceptInvitationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor account created",
                        "schema": {
                            "$ref": "#/definitions/users.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Invitation already used or revoked",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "410": {
                        "description": "Invalid or expired token",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "clinics.InvitationResponse": {
            "type": "object",
            "properties": {
                "accepted_at": {
                    "type": "string"
                },
                "clinic_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "doctor@hospital.com"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440003"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                }
            }
        },
        "diagnoses.CreateDiagnosisRequest": {
            "type": "object",
            "required": [
                "diagnosis_text",
                "diagnosis_type"
            ],
            "properties": {
                "diagnosis_text": {
                    "type": "string"
                },
                "diagnosis_type": {
                    "type": "string",
                    "enum": [
                        "provisional",
                        "final"
                    ]
                },
                "icd_code": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
        "diagnoses.Diagnosis": {
            "type": "object",
            "properties": {
                "diagnosed_at": {
                    "type": "string"
                },
                "diagnosis_text": {
                    "type": "string"
                },
                "diagnosis_type": {
                    "type": "string"
                },
                "encounter_id": {
                    "type": "string"
                },
                "icd_code": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
...
        "invitations.AcceptInvitationRequest": {
            "type": "object",
            "required": [
                "first_name",
                "last_name",
                "license_number",
                "password",
                "phone",
                "specialization"
            ],
            "properties": {
                "first_name": {
                    "description": "Invited doctor's first name",
                    "type": "string",
                    "example": "Jane"
                },
                "last_name": {
                    "description": "Invited doctor's last name",
                    "type": "string",
                    "example": "Doe"
                },
                "license_number": {
                    "description": "Medical license number",
                    "type": "string",
                    "example": "LIC-12345"
                },
                "password": {
                    "description": "Desired account password (min 8 characters)",
                    "type": "string",
                    "minLength": 8,
                    "example": "SecurePass789!"
                },
                "phone": {
                    "description": "Invited doctor's phone number",
                    "type": "string",
                    "example": "+254712345678"
                },
                "specialization": {
                    "description": "Specialization area (e.g. Pediatrics, Cardiology)",
                    "type": "string",
                    "example": "Pediatrics"
                }
            }
        },
        "invitations.CreateInvitationRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "description": "Email address to invite as a doctor",
                    "type": "string",
                    "example": "doctor@hospital.com"
                }
            }
        },
...
        "response.DataEnvelope": {
            "description": "Standard API success response",
            "type": "object",
            "properties": {
                "data": {
                    "description": "The response payload - structure varies per endpoint"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                }
            }
        },
        "response.MessageBody": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
        "response.MessageEnvelope": {
            "description": "Simple message response",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Human-readable status message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.MessageBody"
                        }
                    ]
                }
            }
        },
...
        "users.DoctorStatus": {
            "type": "string",
            "enum": [
                "active",
                "deactivated"
            ],
            "x-enum-varnames": [
                "DoctorStatusActive",
                "DoctorStatusDeactivated"
            ]
        },
        "users.Role": {
            "type": "string",
            "enum": [
                "patient",
                "doctor",
                "clinic_admin",
                "super_admin"
            ],
            "x-enum-varnames": [
                "RolePatient",
                "RoleDoctor",
                "RoleClinicAdmin",
                "RoleSuperAdmin"
            ]
        },
...
        "users.UserResponse": {
            "type": "object",
            "properties": {
                "blood_type": {
                    "type": "string",
                    "example": "O+"
                },
                "clinic_id": {
                    "description": "Doctor specific fields (nullable)",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "description": "Patient specific fields (nullable)",
                    "type": "string"
                },
                "doctor_status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/users.DoctorStatus"
                        }
                    ],
                    "example": "active"
                },
                "email": {
                    "type": "string",
                    "example": "jane.doe@example.com"
                },
                "emergency_contact_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "emergency_contact_phone": {
                    "type": "string",
                    "example": "+254700000000"
                },
                "first_name": {
                    "type": "string",
                    "example": "Jane"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "license_number": {
                    "type": "string",
                    "example": "LIC-12345"
                },
                "phone": {
                    "type": "string",
                    "example": "+254712345678"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/users.Role"
                        }
                    ],
                    "example": "patient"
                },
                "sex": {
                    "type": "string",
                    "example": "female"
                },
                "specialization": {
                    "type": "string",
                    "example": "Cardiology"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token. Format: **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}